+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/metric/protocol/ProtocolMetricImpl.java
@@ -8,6 +8,8 @@ import com.yahoo.bard.webservice.data.metric.LogicalMetricInfo;
 import com.yahoo.bard.webservice.data.metric.TemplateDruidQuery;
 import com.yahoo.bard.webservice.data.metric.mappers.ResultSetMapper;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
@@ -20,13 +22,18 @@ public class ProtocolMetricImpl extends LogicalMetricImpl implements ProtocolMet
 
     protected final ProtocolSupport protocolSupport;
 
+    protected List<LogicalMetric> dependentMetrics;
+
     /**
      * Constructor.
      *
      * @param logicalMetricInfo  The metadata for the metric
      * @param templateDruidQuery  Query the metric needs
      * @param calculation  Mapper for the metric
+     *
+     * @deprecated User {@link #ProtocolMetricImpl(LogicalMetricInfo, TemplateDruidQuery, ResultSetMapper, List)}
      */
+    @Deprecated
     public ProtocolMetricImpl(
             @NotNull LogicalMetricInfo logicalMetricInfo,
             @NotNull TemplateDruidQuery templateDruidQuery,
@@ -36,7 +43,8 @@ public class ProtocolMetricImpl extends LogicalMetricImpl implements ProtocolMet
                 logicalMetricInfo,
                 templateDruidQuery,
                 calculation,
-                DefaultSystemMetricProtocols.getStandardProtocolSupport()
+                DefaultSystemMetricProtocols.getStandardProtocolSupport(),
+                Collections.emptyList()
         );
     }
 
@@ -47,16 +55,64 @@ public class ProtocolMetricImpl extends LogicalMetricImpl implements ProtocolMet
      * @param templateDruidQuery  Query the metric needs
      * @param calculation  Mapper for the metric
      * @param protocolSupport  A identify and return protocols supported for this metric.
+     *
+     * @deprecated Use
+     * {@link #ProtocolMetricImpl(LogicalMetricInfo, TemplateDruidQuery, ResultSetMapper, ProtocolSupport, List)}
      */
+    @Deprecated
     public ProtocolMetricImpl(
             @NotNull LogicalMetricInfo logicalMetricInfo,
             @NotNull TemplateDruidQuery templateDruidQuery,
             ResultSetMapper calculation,
             ProtocolSupport protocolSupport
     ) {
+        this(logicalMetricInfo, templateDruidQuery, calculation, protocolSupport, Collections.emptyList());
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param logicalMetricInfo  The metadata for the metric
+     * @param templateDruidQuery  Query the metric needs
+     * @param calculation  Mapper for the metric
+     * @param dependentMetrics Metrics from which this metric depend
+     */
+    public ProtocolMetricImpl(
+            @NotNull LogicalMetricInfo logicalMetricInfo,
+            @NotNull TemplateDruidQuery templateDruidQuery,
+            ResultSetMapper calculation,
+            List<LogicalMetric> dependentMetrics
+    ) {
+        this(
+                logicalMetricInfo,
+                templateDruidQuery,
+                calculation,
+                DefaultSystemMetricProtocols.getStandardProtocolSupport(),
+                dependentMetrics
+        );
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param logicalMetricInfo  The metadata for the metric
+     * @param templateDruidQuery  Query the metric needs
+     * @param calculation  Mapper for the metric
+     * @param protocolSupport  A identify and return protocols supported for this metric.
+     * @param dependentMetrics Metrics which were used to create this metric.
+     */
+    public ProtocolMetricImpl(
+            @NotNull LogicalMetricInfo logicalMetricInfo,
+            @NotNull TemplateDruidQuery templateDruidQuery,
+            ResultSetMapper calculation,
+            ProtocolSupport protocolSupport,
+            List<LogicalMetric> dependentMetrics
+    ) {
         super(logicalMetricInfo, templateDruidQuery, calculation);
         this.protocolSupport = protocolSupport;
+        this.dependentMetrics = dependentMetrics;
     }
+
 
     @Override
     public boolean accepts(String protocolName) {
@@ -75,6 +131,10 @@ public class ProtocolMetricImpl extends LogicalMetricImpl implements ProtocolMet
         return protocolSupport;
     }
 
+    @Override
+    public List<LogicalMetric> getDependentMetrics() {
+        return dependentMetrics;
+    }
     /**
      * All subclasses of {@code ProtocolMetricImpl} MUST override this method and return an instance of the subclassed
      * type. Inheriting this implementation on subclasses will cause the subclass typing to be lost!
@@ -87,22 +147,28 @@ public class ProtocolMetricImpl extends LogicalMetricImpl implements ProtocolMet
                 info,
                 renameTemplateDruidQuery(info.getName()),
                 renameResultSetMapper(info.getName()),
-                getProtocolSupport()
+                getProtocolSupport(),
+                getDependentMetrics()
         );
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) { return true; }
-        if (o == null || getClass() != o.getClass()) { return false; }
+        if (!(o instanceof ProtocolMetric)) {
+            return false;
+        }
         if (!super.equals(o)) { return false; }
-        final ProtocolMetricImpl that = (ProtocolMetricImpl) o;
-        return Objects.equals(protocolSupport, that.protocolSupport);
+        final ProtocolMetric that = (ProtocolMetric) o;
+        if (!Objects.equals(this.dependentMetrics, that.getDependentMetrics())) {
+            return false;
+        }
+        return Objects.equals(protocolSupport, that.getProtocolSupport());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), protocolSupport);
+        return Objects.hash(super.hashCode(), protocolSupport, dependentMetrics);
     }
 
     @Override
@@ -110,6 +176,7 @@ public class ProtocolMetricImpl extends LogicalMetricImpl implements ProtocolMet
         return "ProtocolMetricImpl{" +
                 super.toString() +
                 ", protocolSupport=" + protocolSupport +
+                ", dependentMetrics=" + dependentMetrics +
                 '}';
     }
 }