+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/config/metric/makers/ConstantMakerSpec.groovy
@@ -28,7 +28,8 @@ class ConstantMakerSpec extends Specification {
         LogicalMetric expectedMetric = new ProtocolMetricImpl(
             new LogicalMetricInfo(AGGREGATION_NAME),
             constantQuery,
-            new NoOpResultSetMapper()
+            new NoOpResultSetMapper(),
+            []
         )
 
         and: