+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/druid/model/postaggregation/WithPostAggregationsSpec.groovy
@@ -34,7 +34,7 @@ class WithPostAggregationsSpec extends Specification {
         );
     }
 
-    def "Test dependant dimensions resolve from fields"() {
+    def "Test dependent dimensions resolve from fields"() {
         expect:
         postAggregation.dependentDimensions == ([dimension1, dimension2] as Set)
     }